requirements.txt
asgiref==3.5.2
Django==4.1.3
Pillow==9.3.0
sqlparse==0.4.3

djangorestframework~=3.14.0
pytz~=2022.6
slugify~=0.0.1